Refactoring Added remanining uri template
