+++ src/main/kotlin/no/skatteetaten/aurora/cantus/controller/ErrorHandler.kt
@@ -7,7 +7,11 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.context.request.WebRequest
+import org.springframework.web.reactive.function.client.WebClientResponseException
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler
+import reactor.core.publisher.Mono
+import reactor.core.publisher.toMono
+import java.time.Duration
 
 @ControllerAdvice
 class ErrorHandler : ResponseEntityExceptionHandler() {
@@ -35,3 +39,23 @@ class ErrorHandler : ResponseEntityExceptionHandler() {
         return handleExceptionInternal(e, response, headers, httpStatus, request)
     }
 }
+
+fun <T> Mono<T>.blockNonNullAndHandleError(duration: Duration = Duration.ofSeconds(30), sourceSystem: String? = null) =
+    this.switchIfEmpty(SourceSystemException("Empty response", sourceSystem = sourceSystem).toMono())
+        .blockAndHandleError(duration, sourceSystem)!!
+
+fun <T> Mono<T>.blockAndHandleError(duration: Duration = Duration.ofSeconds(30), sourceSystem: String? = null) =
+    this.handleError(sourceSystem).toMono().block(duration)
+
+private fun <T> Mono<T>.handleError(sourceSystem: String?) =
+    this.doOnError {
+        if (it is WebClientResponseException) {
+            throw SourceSystemException(
+                message = "Error in response, status:${it.statusCode} message:${it.statusText}",
+                cause = it,
+                sourceSystem = sourceSystem,
+                code = it.statusCode.name
+            )
+        }
+        throw SourceSystemException("Error response", it)
+    }
